+++ swagger/src/schema.rs
@@ -52,8 +52,8 @@ pub struct Schema {
     pub title: Option<String>,
     #[serde(rename = "multipleOf")]
     pub multiple_of: Option<i64>,
-    pub maximum: Option<i64>,
-    pub minimum: Option<i64>,
+    pub maximum: Option<f64>,
+    pub minimum: Option<f64>,
     #[serde(rename = "exclusiveMaximum")]
     pub exclusive_maximum: Option<String>,
     #[serde(rename = "maxLength")]
